requirements.txt
pandas
pypdf
PyMuPDF==1.24.8
pillow
openpyxl
pyzbar